crypto: Pull out OlmMachine test helpers to a new module
