+++ crates/matrix-sdk-crypto/src/dehydrated_devices.rs
@@ -388,8 +388,9 @@ mod tests {
     };
 
     use crate::{
-        machine::tests::{
-            create_session, get_prepared_machine_test_helper, to_device_requests_to_content,
+        machine::{
+            test_helpers::get_prepared_machine_test_helper,
+            tests::{create_session, to_device_requests_to_content},
         },
         olm::OutboundGroupSession,
         types::{events::ToDeviceEvent, DeviceKeys as DeviceKeysType},
